Addition of visualization file
moved save_simulation_figure, save_simulation movie, get_timeseries_data, update_3d_plot to visualization.py
